scripts/requirements.txt
PyGithub==2.1.1
requests==2.31.0
googletrans==4.0.0rc1
pyyaml==6.0.1

